rename snake camel case and update use of effected paths
