+++ lua_helpers/astJsonToCode.luau
@@ -1,5 +1,5 @@
 local fs = require("@lute/fs")
-local printer = require("./temp_vendor/lute_printer") -- temp vendoring printer until my updates to support printStatement are released
+local printer = require("./temp_vendor/lutePrinter") -- temp vendoring printer until my updates to support printStatement are released
 local json = require("./json")
 local process = require("@lute/process")
