+++ lua_tests/helpers/typeAnnotationTestHelpers.lua
@@ -2,7 +2,7 @@ local luau = require("@lute/luau")
 local visitor = require("@std/syntax/visitor")
 
 local typeAnnotationVisitor = visitor.createVisitor()
-local e2ecases = require("./ast_json_to_code_helpers").testCases.e2eCases
+local e2ecases = require("./astJsonToCodeHelpers").testCases.e2eCases
 
 --[[
     In writing these tests, I realized we could have just used visitor to annotate types in the first place... (perhaps revisit later)
